Refactor build script to include building PharmGKB databases
The build script has been modified to include the building of the PharmGKB databases for both hg19 and hg38. If the databases do not exist, they are removed and rebuilt using the `build_pharmgkb.sh` script. The existence of the databases is checked at the end of the script to ensure successful creation.