Refactor of Dockerfile and build scripts
The container build scripts have been upgraded to support a multi-stage build process. In the first stage, all databases are downloaded and pre-processed to get their latest version. In the second stage, the container used for the app is built with the data gathered from the first step.
In this phase, a versions.txt file is created to keep track of all the release and download date of all databases stored within the container.